# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Booking API",
        "description": "API documentation for the Travel Booking system",
        "contact": {
            "email": "hey@api-travel.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "List all bookings",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create a new booking",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "flight_number",
                                    "origin",
                                    "destination",
                                    "travel_date",
                                    "departure_time",
                                    "arrival_time",
                                    "seat_class",
                                    "total_price"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Ali Rezaei"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "example": "ali@test.com"
                                    },
                                    "flight_number": {
                                        "type": "string",
                                        "example": "IR455"
                                    },
                                    "origin": {
                                        "type": "string",
                                        "example": "THR"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "DXB"
                                    },
                                    "travel_date": {
                                        "type": "string",
                                        "example": "2026-08-01"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "example": "08:30"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "seat_class": {
                                        "type": "string",
                                        "enum": [
                                            "economy",
                                            "business"
                                        ]
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "example": 2500000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get a booking",
                "operationId": "65258f4a04829afc503334b30421ad4f",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking details"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Update a booking",
                "operationId": "3c2a7a71f0168b95dc25ae8727e44d31",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Ali Rezaei"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "example": "ali@test.com"
                                    },
                                    "flight_number": {
                                        "type": "string",
                                        "example": "IR455"
                                    },
                                    "origin": {
                                        "type": "string",
                                        "example": "THR"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "DXB"
                                    },
                                    "travel_date": {
                                        "type": "string",
                                        "example": "2026-08-01"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "example": "08:30"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "example": "11:00"
                                    },
                                    "seat_class": {
                                        "type": "string",
                                        "enum": [
                                            "economy",
                                            "business"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{booking}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking",
                "operationId": "a3bcc2bed0e2824d299bb853832c1db3",
                "parameters": [
                    {
                        "name": "booking",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    },
                    "422": {
                        "description": "Already cancelled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}